requirements.txt
opencv-python-headless==4.5.5.62
torchvision==0.11.1
Pillow==9.0.1
PyYAML==6.0
matplotlib==3.5.1
seaborn==0.11.2
pafy==0.5.5
youtube-dl==2020.12.2